+++ Caper/Caper/MapviewMappingUtilities.h
@@ -7,7 +7,7 @@ class MapviewMappingUtilities : public MappingUtilities
 {
 public:	
 	string GetSequence( string & aLine );
-  int GetIndex( string & aLine );    
+  long long GetIndex( string & aLine );    
   string GetName( string & aLine );
   string GetContigIdent( string & aLine );	
   string GetStrand( string & aLine );
